Выполняет ДЗ "Модуляция"
